[TASK] Remove old and deprecated code
dam_ttcontent 1.2 requires TYPO3 4.5+. Therefore some old version
switches can be removed.

A call to t3lib_div::intInRange needs to be changed to the wrapper
method in class.tx_dam.php because these method is deprecated in
TYPO3 4.6.

Version dependency was raised to DAM 1.3 because it depends on
static methods introduced with this version.

Change-Id: I236154dbefbc8d8a4d5501861f860da91ea71d88
Resolves: #18298
Releases: 1.2