[BUGFIX] ctypes_textpic_image_add_orig_field broken
If ctypes_textpic_image_add_orig_field is set in extension manager
both fields are shown in content element form but still only images
from tx_damttcontent_files are rendered.

This is a regression of #M16349 where a variable in ext_conf_template.txt
was changed but the condition in localconf.php not.

Change-Id: I6ac476ac7bcac89b39e9eaaf0a11827f948d7c4f
Resolves: #M18484
Releases: 1.3