[BUGFIX] dam_ttcontent seems not to respect renderMethod
As a results of the changes in css_styled_content, dam_ttcontent 1.3.0
was updated to be able to use the accessibility rendering.

Since that change, multiple pictures with no caption set in the tt_content
field imagecaption were always rendered as ul/li instead of dl/dt, even if
the renderMethod was dl.

This is due to the fact that CSC has a fallbackRendering if certain
conditions (only one single pictures, global caption) are met. This
fallbackRendering is not valid for dam_ttcontent since dam_ttcontent
usually uses captions from the DAM record. Therefore the
fallbackRendering is removed (as it was in 1.2).

Change-Id: I59b6102b4f3fb26cd9e7c707d53dd22f6058bae3
Fixes: #41493
Releases: 1.3